+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
         applicationId "com.snapxeats"
         minSdkVersion 21
         targetSdkVersion 23
-        versionCode 602
+        versionCode 603
         versionName "0.6"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
         resValue "string", "HOCKEYAPP_APP_ID", HOCKEYAPP_APP_key
