+++ app/src/main/java/com/snapxeats/ui/restaurant/RestImagesAdapter.java
@@ -10,7 +10,7 @@ import android.widget.FrameLayout;
 import android.widget.ImageView;
 
 import com.snapxeats.R;
-import com.snapxeats.common.model.restaurantDetails.RestaurantPics;
+import com.snapxeats.common.model.restaurantInfo.RestaurantPics;
 import com.squareup.picasso.Picasso;
 
 import java.util.List;
